+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Class names with CSS Handles
+
 ## [0.10.1] - 2020-04-01
 ### Changed
 - Lock page scroll with `overflow: hidden` instead of fixing its position.